# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bindu API",
        "title": "Bindu API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7000",
    "basePath": "/",
    "paths": {
...
        "/permissions/organisations": {
            "get": {
                "description": "Get all organisation permissions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation_Permissions"
                ],
                "summary": "Show all organisation permissions",
                "operationId": "get-all-org-permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Query",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/organisation.paging"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create organisation permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation_Permissions"
                ],
                "summary": "Create organisation permission",
                "operationId": "add-org-permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permission Object",
                        "name": "Permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organisation.organisationPermission"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.OrganisationPermission"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/permissions/spaces": {
            "get": {
                "description": "Get all Space permissions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space_Permissions"
                ],
                "summary": "Show all Space permissions",
                "operationId": "get-all-space-permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Query",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github.com_factly_bindu-server_action_permission_space.paging"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Space permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space_Permissions"
                ],
                "summary": "Create Space permission",
                "operationId": "add-space-permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permission Object",
                        "name": "Permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/space.spacePermission"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.SpacePermission"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/requests/organisations": {
            "get": {
                "description": "Get all organisation permissions requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation_Permissions_Request"
                ],
                "summary": "Show all organisation permissions requests",
                "operationId": "get-all-org-permissions-requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/organisation.paging"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create organisation permission request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation_Permissions_Request"
                ],
                "summary": "Create organisation permission request",
                "operationId": "add-org-permission-request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Request Object",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/organisation.organisationPermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.OrganisationPermissionRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/requests/spaces": {
            "get": {
                "description": "Get all space permissions requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space_Permissions_Request"
                ],
                "summary": "Show all space permissions requests",
                "operationId": "get-all-space-permissions-requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github.com_factly_bindu-server_action_request_space.paging"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create space permission request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space_Permissions_Request"
                ],
                "summary": "Create space permission request",
                "operationId": "add-space-permission-request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Request Object",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/space.spacePermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.SpacePermissionRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/spaces": {
            "get": {
                "description": "Get all spaces",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space"
                ],
                "summary": "Show all spaces",
                "operationId": "get-all-spaces",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/space.orgWithSpace"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create space",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space"
                ],
                "summary": "Create space",
                "operationId": "add-space",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Space Object",
                        "name": "Space",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/space.space"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Space"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github.com_factly_bindu-server_action_permission_space.paging": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/space.spaceWithPermissions"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github.com_factly_bindu-server_action_request_space.paging": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SpacePermissionRequest"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.OrganisationPermission": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "model.OrganisationPermissionRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Space": {
            "type": "object",
            "properties": {
                "contact_info": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fav_icon": {
                    "$ref": "#/definitions/model.Medium"
                },
                "fav_icon_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "$ref": "#/definitions/model.Medium"
                },
                "logo_id": {
                    "type": "integer"
                },
                "logo_mobile": {
                    "$ref": "#/definitions/model.Medium"
                },
                "logo_mobile_id": {
                    "type": "integer"
                },
                "mobile_icon": {
                    "$ref": "#/definitions/model.Medium"
                },
                "mobile_icon_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "site_address": {
                    "type": "string"
                },
                "site_title": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "social_media_urls": {
                    "type": "string"
                },
                "tag_line": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "verification_codes": {
                    "type": "string"
                }
            }
        },
        "model.SpacePermission": {
            "type": "object",
            "properties": {
                "charts": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "model.SpacePermissionRequest": {
            "type": "object",
            "properties": {
                "charts": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.organisationUser": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "organisation.organisationPermission": {
            "type": "object",
            "required": [
                "organisation_id"
            ],
            "properties": {
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                }
            }
        },
        "organisation.organisationPermissionRequest": {
            "type": "object",
            "required": [
                "organisation_id",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "organisation.paging": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrganisationPermissionRequest"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "space.orgWithSpace": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "permission": {
                    "$ref": "#/definitions/model.organisationUser"
                },
                "slug": {
                    "type": "string"
                },
                "spaces": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Space"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "space.space": {
            "type": "object",
            "required": [
                "name",
                "organisation_id"
            ],
            "properties": {
                "contact_info": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fav_icon_id": {
                    "type": "integer"
                },
                "logo_id": {
                    "type": "integer"
                },
                "logo_mobile_id": {
                    "type": "integer"
                },
                "mobile_icon_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "site_address": {
                    "type": "string"
                },
                "site_title": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "social_media_urls": {
                    "type": "string"
                },
                "tag_line": {
                    "type": "string"
                },
                "verification_codes": {
                    "type": "string"
                }
            }
        },
        "space.spacePermission": {
            "type": "object",
            "required": [
                "space_id"
            ],
            "properties": {
                "charts": {
                    "type": "integer"
                },
                "space_id": {
                    "type": "integer"
                }
            }
        },
        "space.spacePermissionRequest": {
            "type": "object",
            "required": [
                "space_id",
                "title"
            ],
            "properties": {
                "charts": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "space_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}